requirements.txt
asgiref==3.2.10
awsebcli==3.20.10
botocore==1.31.85
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.3
Django==3.1
django-admin-thumbnails==0.2.9
django-environ==0.11.2
idna==3.4
isort==5.12.0
jmespath==1.0.1
mysqlclient==2.2.4
pathspec==0.10.1
Pillow==9.5.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3
pywin32==306
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
urllib3==1.26.18
wcwidth==0.1.9
